Add files via upload
This is the complete folder containing all the relevant documents, figures and codes for the CHEME 7770 FINAL PROJECT.

Inside can be found the SUPPLEMENTARY MATERIALS folder which is referenced in the paper. It contains all graphs and figures output from the code along with the visual simulation movies and code. If you would like to run the processing code yourself, run the simple elastic Monte Carlo MATLAB simulation in the same directory as the Processing code. Be sure to uncomment the file generation section of the code. The code will output numerous files which is somewhat inelegant these are used by the Processing code. The Processing code can run visual simulations at all stiffnesses by changing the s value.

The powerpoint presentation and final paper can be found in the main folder along with all the codes including the python, original Odde et al code, and the modified simple elastic and viscoelastic MATLAB codes.